requirements.txt
cachetools==2.1.0
certifi==2018.10.15
chardet==3.0.4
google-api-core==1.5.1
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-vision==0.34.0
googleapis-common-protos==1.5.5
grpcio==1.16.0
httplib2==0.11.3
idna==2.7
nltk==3.3
oauth2client==4.1.3
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pytz==2018.7
redis==2.10.6
requests==2.20.0
rsa==4.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.24
